translate/2to3: Honor unit.Enable bool, translate to unit.Enabled
Adds translation of the spec v2.x unit.Enable field to unit.Enabled in
spec v3.x.

The Enabled field wins over Enable, since Enable is deprecated in
spec v2 and removed in v3. It does so following the apparent intent
of the upstream code [1] which actually does the opposite for
Enable=true Enabled=false
because the first matching line in a systemd preset wins.

unit.Enabled can be omitted from the config when it is not explicitly
set.

Setting unit.Enable to false has no effect.

[1] https://github.com/coreos/ignition/blob/b4d18ad3fcb278a890327f858c1c10256ab6ee9d/internal/exec/stages/files/units.go#L32