translate/2to3: Deduplicate translated files and units
This commit adds functions to ensure the emitted spec v3 config
does not contain duplicate file and unit entries.

In spec v2 duplicate files and units are allowed, with the last one
defined overriding previous ones.
That is however forbidden in spec v3.

Until now the converter has not checked for duplicates in the spec v3
output.

For duplicate files, the last one defined wins and is written.
For duplicate units, the last one with non-empty contents wins and is written.
For duplicate unit dropins, the last one wins and is written.

In the case that a duplicate unit defines dropins only and has no
content, the dropins are appended to the previous unit's dropins,
replacing dropins of the same name if necessary.