translate/v2*tov3*: Add RemoveDuplicateFilesAndUnits helper func
This was originally written by Jerry Zhang for inclusion in the MCO [1].
Adding it here as the deduplication step often times has to be performed
in preparation of translation when dealing with v2 config in order to
sanitize it.

RemoveDuplicateFilesAndUnits is a helper function that removes duplicated files/units from
spec v2 config, since neither spec v3 nor the translator function allow for duplicate file
entries in the config.
This functionality is not included in the Translate function, but may be useful in cases
where configuration has to be sanitized before translation.
For duplicates, it takes ordering into consideration by taking the file/unit contents from
the slice with the highest index value, which is assumed to be the latest revision.
Multiple filesystems are not considered, it is assumed there is only one.
File appendices are not considered.
Unit dropins are concat'ed, i.e. appended if no duplicate dropin of the same name exists.

[1] https://github.com/openshift/machine-config-operator/commit/12668c044cbc5d4d7c2e3c90cd6ab13fa7d6f49d

Co-authored-by: Yu Qi Zhang <jerzhang@redhat.com>